Finalised and reorganised code back into made folder
