--- recognition/VQVAE_James_Lewis/code/dataset.py
@@ -1,102 +1,0 @@
-import numpy as np
-import nibabel as nib
-from tqdm import tqdm
-from skimage.transform import resize
-
-def to_channels(arr: np.ndarray, dtype=np.uint8) -> np.ndarray:
-    channels = np.unique(arr)
-    res = np.zeros(arr.shape + (len(channels),), dtype=dtype)
-    for c in channels:
-        c = int(c)
-        res[..., c:c + 1][arr == c] = 1
-    return res
-
-
-# Load medical image functions
-def load_data_2D(imageNames, normImage=False, categorical=False, dtype=np.float32,
-                 getAffines=False, early_stop=False):
-    '''
-    Load medical image data from names, cases list provided into a list for each.
-
-    This function pre-allocates 4D arrays for conv2d to avoid excessive memory usage.
-
-    normImage : bool (normalise the image 0.0 -1.0)
-    early_stop : Stop loading pre-maturely, leaves arrays mostly empty, for quick loading and testing scripts.
-    '''
-    affines = []
-
-    # Get fixed size
-    num = len(imageNames)
-    first_case = nib.load(imageNames[0]).get_fdata(caching='unchanged')
-
-    if len(first_case.shape) == 3:
-        first_case = first_case[:, :, 0]  # Sometimes extra dims, remove
-
-    if categorical:
-        first_case = to_channels(first_case, dtype=dtype)
-        rows, cols, channels = first_case.shape
-        images = np.zeros((num, rows, cols, channels), dtype=dtype)
-    else:
-        rows, cols = first_case.shape
-        images = np.zeros((num, rows, cols), dtype=dtype)
-
-    for i, inName in enumerate(tqdm(imageNames)):
-        niftiImage = nib.load(inName)
-        inImage = niftiImage.get_fdata(caching='unchanged')  # Read disk only
-        affine = niftiImage.affine
-
-        if len(inImage.shape) == 3:
-            inImage = inImage[:, :, 0]  # Sometimes extra dims in HipMRI_study data
-
-        inImage = resize(inImage, (256, 128), preserve_range=True, anti_aliasing=True).astype(dtype)
-
-        if normImage:
-            # ~ inImage = inImage / np.linalg.norm(inImage)
-            # ~ inImage = 255. * inImage / inImage.max()
-            inImage = (inImage - inImage.mean()) / inImage.std()
-
-        if categorical:
-            inImage = to_channels(inImage, dtype=dtype)
-            images[i, :, :, :] = inImage
-        else:
-            images[i, :, :] = inImage
-
-        affines.append(affine)
-
-        if i > 20 and early_stop:
-            break
-
-    if getAffines:
-        return images, affines
-    else:
-        return images
-
-class DataLoader:
-    def __init__(self, dataset, batch_size=32, shuffle=True):
-        self.dataset = dataset
-        self.batch_size = batch_size
-        self.shuffle = shuffle
-        self.indices = np.arange(len(dataset))
-        self.current_index = 0
-
-        if self.shuffle:
-            np.random.shuffle(self.indices)
-    def __len__(self):
-        return len(self.dataset) // self.batch_size
-    def __iter__(self):
-        self.current_index = 0
-        if self.shuffle:
-            np.random.shuffle(self.indices)
-        return self
-
-    def __next__(self):
-        if self.current_index >= len(self.indices):
-            raise StopIteration
-
-        start_idx = self.current_index
-        end_idx = min(start_idx + self.batch_size, len(self.indices))
-        batch_indices = self.indices[start_idx:end_idx]
-
-        batch_data = [self.dataset[i] for i in batch_indices]
-        self.current_index = end_idx
-        return np.array(batch_data)
